Fix PHPUnit test running
We don't want a PHPUnit.xml file in our module because we're fine with
the defaults that the CI comes up with. However, we do need some
settings to properly run them locally so we pass those in through our
composer command.

Apparently Drupal also requires `behat/mink` for the PHPUnit test-suite
so we add the dev dependency.