Add mariadb client to web container
This is needed for some drush commands such as site-install to work
correctly. The mariadb client is also aliased to `mysql` to make that
work.

We move around the xdebug adding since it's relatively slow and this
makes it possible to add new quickly installed tools without rebuilding
those parts of the image.