Add Behat test set-up
PHPUnit is not a great tool to write UI tests with. Drupal core has
Nightwatch but that's also not a tool the project maintainer is familiar
with.

Behat is a tool the maintainer is familiar with and allows us to
describe scenario's in human readable texts that can be discussed in
issues.

We add a Behat config that uses Chrome as well as some useful helpers to
get debug output in case of test failures as well as a way to easily
interface with Drupal.

The dependencies needed to run these tests are not added in the module
since some more infrastructure is needed. Instead those were added to
the development environment at https://github.com/Kingdutch/rtseo_dev