Issue #3204291 by MarcKwee, StryKaizer, batigolix, medwassim, BramDriesen: Call to undefined method Drupal\taxonomy\Entity\Term::getTitle()
This replaces the getTitle call which is specific to nodes with `label`.
Under the hood that is equivalent to calling `getTitle` for nodes but
for Term's the entity system ensures it's equivalent to callling
`getName`. By calling the `label` method we make use of the
`EntityInterface` which we already know we have so we don't need to
guess at methods using `method_exists`.