Issue #3122871 by crmn, Ruuds, BramDriesen: Analyze frontend theme and view mode output
This commit falls back to the default front-end theme by default, even
if content is edited within the context of the admin theme. It also
provides a per-form configuration to change this.

While we're doing the plumbing for the theme we do a similar plumbing
for the view mode which was an open todo. The default view mode is
changed from `full` to `default` since that's the Drupal default view
mode and `full` is not always available.