Update test setup to monorepo and fix tests
This reconfigures our testing tools to work with the new monorepo
set-up. We also previously updated some testing tools and as a result we
update some of the files stylistically to make sure all those tools are
happy again.

The one FunctionalBrowser test gets rewritten to Behat because I can
never be bothered to set-up PHPUnit's browser based testing and that's
what I have Behat for.