Add a GitHub workflow for testing
This adds a GitHub workflow that runs PHPCS, PHPStan, PHPUnit and Behat.
This ensures that we can continuously check the quality for every commit
and PR and not forget about it.

It can be improved in the future by adding better integration with the
reporting for GitHub, but for now we're content with just running the
checks.